+++ sleeper_api/__init__.py
@@ -55,4 +55,4 @@ __all__ = [
     "UserNotFoundError",
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"